// src/app/cases/models/cases-type.ts
export type CasesType =
  'ALL' |
  'OFFICIAL' |
  'DAILY' |
  'STREETWEAR' |
  'ELECTRONICS' |
  'ACCESSORIES' |
  'TOP 100' |
  'NEW CASE' |
  'GAMING' |
  'FREE';

export enum CaseType {
  all = 'ALL',
  official = 'OFFICIAL',
  daily = 'DAILY',
  streetwear = 'STREETWEAR',
  electronics = 'ELECTRONICS',
  accessories = 'ACCESSORIES',
  top = 'TOP 100',
  new = 'NEW CASE',
  cent = 'GAMING',
  free = 'FREE'
}